+++ MathEvaluation/Entities/MathEntity.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MathEvaluation.Context;
+namespace MathEvaluation.Entities;
 
 /// <summary>
 /// Base class for a math entity.
@@ -15,7 +15,7 @@ public abstract class MathEntity : IMathEntity
 
     /// <summary>Initializes a new instance of the <see cref="MathEntity" /> class.</summary>
     /// <param name="key">The key.</param>
-    /// <exception cref="System.ArgumentNullException">key</exception>
+    /// <exception cref="ArgumentNullException">key</exception>
     protected MathEntity(string? key)
     {
         Key = key ?? throw new ArgumentNullException(nameof(key));
