+++ MathEvaluation/Entities/MathGetValueFunction.cs
@@ -1,6 +1,7 @@
 ﻿using System;
+using MathEvaluation.Context;
 
-namespace MathEvaluation.Context;
+namespace MathEvaluation.Entities;
 
 /// <summary>
 /// The getting value function.
@@ -19,7 +20,7 @@ public class MathGetValueFunction<T> : MathEntity
     /// <summary>Initializes a new instance of the <see cref="MathGetValueFunction{T}" /> class.</summary>
     /// <param name="key">The key.</param>
     /// <param name="fn">The getting value function.</param>
-    /// <exception cref="System.ArgumentNullException">fn</exception>
+    /// <exception cref="ArgumentNullException">fn</exception>
     public MathGetValueFunction(string? key, Func<T> fn)
         : base(key)
     {
