+++ MathEvaluation/Entities/MathOperandsOperator.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace MathEvaluation.Context;
+namespace MathEvaluation.Entities;
 
 /// <summary>
 /// The math operator processes the left and right math operands.
@@ -20,7 +20,7 @@ public class MathOperandsOperator<T> : MathEntity
     /// <param name="key">The key (the operator notation).</param>
     /// <param name="fn">The function.</param>
     /// <param name="precedece">The operator precedence.</param>
-    /// <exception cref="System.ArgumentNullException">fn</exception>
+    /// <exception cref="ArgumentNullException">fn</exception>
     public MathOperandsOperator(string? key, Func<T, T, T> fn, int precedece)
         : base(key)
     {
