+++ MathEvaluation/Entities/MathOperator.cs
@@ -1,9 +1,10 @@
 ﻿using System;
+using MathEvaluation.Context;
 
-namespace MathEvaluation.Context;
+namespace MathEvaluation.Entities;
 
 /// <summary>
-/// The math operator can process the left and right expressions.
+/// The math operator processes the left and right expressions.
 /// </summary>
 /// <typeparam name="T"></typeparam>
 public class MathOperator<T> : MathEntity
@@ -20,7 +21,7 @@ public class MathOperator<T> : MathEntity
     /// <param name="key">The key (the operator notation).</param>
     /// <param name="fn">The function.</param>
     /// <param name="precedece">The operator precedence.</param>
-    /// <exception cref="System.ArgumentNullException">fn</exception>
+    /// <exception cref="ArgumentNullException">fn</exception>
     public MathOperator(string? key, Func<T, T, T> fn, int precedece = (int)EvalPrecedence.Basic)
         : base(key)
     {
