+++ MathEvaluation/MathEvaluator.Boolean.cs
